add common function to extract frame_range
