+++ Sports2D/Utilities/skeletons.py
@@ -30,7 +30,7 @@
 '''
 
 ## INIT
-from anytree import Node, RenderTree, PreOrderIter
+from anytree import Node
 
 
 ## AUTHORSHIP INFORMATION
